requirements.txt
streamlit>=1.29
python-dotenv>=1.0
openai>=1.7           # official SDK (needed by langchain-openai)
langchain-core>=0.1.14
langchain>=0.1.11
langchain-community>=0.0.21
langchain-openai>=0.0.5
langchain-groq>=0.0.2
faiss-cpu>=1.7